Created 2D Unity Project!
This file is the continuation of the Zombie Shooter Project. This project will contain my changes and the overall development progress of my game!

List of Changes:
Zombie Shooter Project renamed to 2D-Unity-Project (This will change when a name has been decided!)

-----------------------------------------

List of Fixes:
Fixed the score counter - When the player killed zombies, 10 points was supposed to be added to the Score counter in the top left. This has been fixed and points are now rewarded!

Fixed the blood explosion - When the player's bullets collided with zombies, no blood appeared, this has been fixed and the blood explosion now displays correctly!